Move feature-oriented helpers to appropriate location
